Migrate 'value' test
